requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.10.0
bidict==0.22.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
click==7.1.2
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
dnspython==2.3.0
dominate==2.7.0
exceptiongroup==1.0.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
Flask-SocketIO==5.3.1
Flask-uWSGI-WebSocket==0.6.1
fonttools==4.38.0
fpdf==1.7.2
h11==0.14.0
idna==2.5
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
macholib==1.16
MarkupSafe==2.1.1
matplotlib==3.6.2
minimalmodbus==2.0.1
numpy==1.23.4
opencv-python==4.6.0.66
outcome==1.2.0
packaging==22.0
Pillow==9.3.0
pycparser==2.21
pyinstall==0.1.4
pyModbusTCP==0.2.0
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt6==6.4.0
PyQt6-Charts==6.4.0
PyQt6-Charts-Qt6==6.4.1
pyqt6-plugins==6.1.0.2.2
PyQt6-Qt6==6.4.1
PyQt6-sip==13.4.0
pyqt6-tools==6.1.0.3.2
PyQt6-WebEngine==6.4.0
PyQt6-WebEngine-Qt6==6.4.1
PyQtChart==5.15.6
PyQtChart-Qt5==5.15.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
PySide6==6.4.0.1
PySide6-Addons==6.4.0.1
PySide6-Essentials==6.4.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.1
qt6-applications==6.1.0.2.2
qt6-tools==6.1.0.1.2
requests2==2.16.0
selenium==4.5.0
shiboken6==6.4.0.1
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
uWSGI==2.0.20
visitor==0.1.3
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
